feat: add separate network for mlbvip
This is not working at all, might abandon. Traffic not flowing from
dummy port for VIP to the node that is actually advertising IP. Not sure
what could be misconfigured. Read into allowed address pairs but that
should not be needed with port security disabled from what I can tell.