OSPC-541 Deploy LetsEncrypt cluster issuer
This introduces a required variable `letsencypt_email`. An email address is
required for the lets encrypt cluster issuer.  If you do not specify the
variable when running the playbook then you will be prompted for the email
in the middle of the playbook run. Which is not great.

I suggest you provide the variable on the commandline so that you don't get
prompted in the middle of the playbook run.  Ex:

'''
ansible-playbook ~/genestack_post_deploy.yaml -e "letsencrypt_email=<valid email address>"
'''