feature: Add general flex-node module
Adds a new module, flex-node, for reuse in creating flex nodes. There was a lot
of repetition in creating nodes for the different roles. Using the flex-node
module with appropriate variables facilitates less replication.

The following roles have been converted to using the flex-node module:
- kubernetes
- controller
- worker
- compute
- network

Storage roles still need to be converted.

Signed-off-by: Chris Blumentritt <chris.blumentritt@rackspace.com>